# api/openapi.json
{
  "components": {
    "schemas": {
...
      "dto.CommentResponse": {
        "properties": {
          "author": {
            "$ref": "#/components/schemas/dto.UserSummary"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "item_id": {
            "type": "integer"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "dto.CommentUpdateRequest": {
        "properties": {
          "content": {
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "type": "object"
      },
...
      "dto.UserSummary": {
        "properties": {
          "avatar_url": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "last_name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "handlers.ErrorResponse": {
        "properties": {
          "code": {
            "description": "Machine-readable error code",
            "type": "string"
          },
          "details": {
            "description": "Additional error details"
          },
          "error": {
            "description": "Human-readable message",
            "type": "string"
          },
          "request_id": {
            "description": "Request correlation ID",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "API bearer token in the form `Bearer crw_*`. Token scopes are checked per route.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Public REST API for Windshift work-management. Authenticate every request with a bearer token (`Authorization: Bearer crw_*`). Tokens are scoped per route — see each operation's `security` block for the required scopes.",
    "title": "Windshift REST API",
    "version": "1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/comments/{id}": {
      "delete": {
        "description": "The caller must be the comment author or hold edit permission on the comment's workspace.",
        "parameters": [
          {
            "description": "Comment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Invalid comment ID"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Token lacks the items:delete scope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Comment not found or caller cannot delete it"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete a comment",
        "tags": [
          "comments"
        ]
      },
      "get": {
        "description": "Returns 404 (not 403) when the caller cannot view the comment's parent workspace — workspace existence is never leaked.",
        "parameters": [
          {
            "description": "Comment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.CommentResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Invalid comment ID"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Token lacks the items:read scope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Comment not found or not visible to caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get a comment by ID",
        "tags": [
          "comments"
        ]
      },
      "put": {
        "description": "The caller must be the comment author or hold edit permission on the comment's workspace.",
        "parameters": [
          {
            "description": "Comment ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/dto.CommentUpdateRequest"
              }
            }
          },
          "description": "Updated content",
          "required": true,
          "x-originalParamName": "body"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/dto.CommentResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Invalid comment ID, request body, or empty content"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Token lacks the items:write scope"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Comment not found or caller cannot edit it"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update a comment",
        "tags": [
          "comments"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/rest/api/v1"
    }
  ]
}